+++ src/Controller/Api/PluginConfigController.php
@@ -208,7 +208,7 @@ class PluginConfigController extends AbstractController
             }
         }
 
-        if (!$credentials['endpoint']) {
+        if (empty($credentials['endpoint'])) {
             $credentials['endpoint'] = $mode === 'live' ? WorldlineSDKAdapter::LIVE_ENDPOINT : WorldlineSDKAdapter::TEST_ENDPOINT;
         }
         return $credentials;