+++ src/Service/OldOrderProcessor.php
@@ -135,7 +135,7 @@ class OldOrderProcessor
         }
 
         if ($timeInterval > 0) {
-            $qb->andWhere("UNIX_TIMESTAMP(NOW()) - UNIX_TIMESTAMP(o.updated_at) > :timeInterval")
+            $qb->andWhere("TIMESTAMPDIFF(SECOND, o.updated_at, UTC_TIMESTAMP()) > :timeInterval")
                 ->setParameter('timeInterval', $timeInterval);
         }
         return $qb->execute()->fetchAllAssociative() ?: [];